Rename button.jsx to randomize.jsx
